Multiplayer
Turns out what I had before this was a very very shaky foundation.
A lot of foundation had to be built or rebuilt. I can only hope that
this foundation I've built to get multiplayer working is good enough
for a good amount of time.

A big addition to the foundation is the concept of "snapshots", or
what the game looks like from a neutral observer.

I've also started moving JS modules out into their own files when
they start getting too big. I also generated a general layer system
for the JS application. Turns out what I have in App.js needs to be
split out into four main parts:

1. Main function that wires all the parts together.
2. DOM speaking layer. This layer knows the DOM and the View,
   gathering events from the browser and calling various functions
   in the View. This is the only layer that knows about the browser.
3. View layer which holds a synthetic version of the components on
   the page.
4. Controller which takes events from the View and Channel, updates
   the Gald model, and otherwise acts as a dispatcher.

The Gald Model is structured such that it delegates to an inner
model based on the status (Lobby|Play|Over) what it should do
when it gets a message.