Auto-updating lobby; multiple JS targets
The logic for the auto-updating lobby - the visible feature change,
was actually the easiest part of this PR. It consists of the lobby
channel, race manager, and lobby/main.js, and user_socket files.

Basically, there's a new channel "lobby" (no subchannels), which
does *not* currently handle any messages from the client. Instead,
when a race is created or destroyed, that is broadcasted to anybody
on the lobby page, which updates the list dynamically.

If, for whatever reason, JS is disabled by the point a user gets to
the lobby, it'll still show the lobby at the time of the request. Go
figure to whoever will find that useful.

--

The harder part was making multiple JS targets. I *don't* want
people downloading the entire game on their initial visit to the site
unless they are actually going to a game.

So, instead of a single target, here's what the design is:

* One target per actual page.
* A 'util' target that holds common utilities. Things like the
  phoenix channel stuff, but also r-result whenever I port it over.
* A 'vender' target that holds vender JS - whenever I actually get some.

I cannot seem to make compilation targets rely on each other, so I'm sort
of stuck in a dynamic library loading situation where it's up to each HTML
page to ensure they are loaded. This makes me sad, but I can live with it.

Anyways, this message is getting pretty long. There's a new minor feature
and another part of the site's foundation complete. There's still a lot
of foundation to go though.