Split game 'url' and 'name' into two separate concepts.
The 'url' is managed by the GaldSite.RaceManager. The name is
managed by Gald itself.

To avoid working with randomness as long as possible, I'm temporarily
designing the URL to be made with regexp. In the future, it'll be
a random string (probably an 8 character long base 64 string)

I renamed the race/index template to race/lobby.