Event outputting, turn based Gald race.
Addressability was solved. Everything holds a reference to the
supervisor and the components are globally named using the
supervisor.

There is now an output stream for the game and one for each player.

Right now the player output stream receives no messages, but the
game receives everything.

I set up a system where there's a round that turns turns which
starts screen sequences. Right now the only screen sequence is
dice rolling. Wheee...

What /was/ Gald.Race was renamed to Gald.Controller and the
supervisor is now at Gald.Supervisor. No longer does it create
both the controller and supervisor under a supervisor though,
the supervisor just starts the controller first, along with
other processes.

Victory conditions are also checked in a Victory module server.